Analyse the code as NanoFaaS does, and fix what the analysis reports
SpotBugs 4.10.4 and Error Prone 2.50.0 run with NanoFaaS's settings: every
finding is reported, none fails the build, and SpotBugs leaves the test source
sets alone. JaCoCo moves to 0.8.14 and gains NanoFaaS's 85% instruction rule,
run on demand through jacocoTestCoverageVerification rather than by check.

One difference: this repository compiles thousands of generated protobuf classes
into main. Error Prone skips them by path (they carry no @Generated it
recognises), SpotBugs by package, and the coverage rule uses the exclusion the
coverage report already had.

What the analysis found, now fixed:

- Container and Image copied their labels only when ProtoMapper built them; the
  records now copy them themselves, as NetworkAttachment and ExecSpec already do.
- ContainerJournal caught NullPointerException to detect a record with no
  container; it now checks for the missing entry and says so.
- Three Path.getFileName() and getParent() results used without a null check
  are replaced by forms that cannot be null.
- toLowerCase() without a Locale, an implicit long-to-double conversion, and
  getBytes() with the platform charset in tests are made explicit.
- Futures ignored on purpose are either not created (execute instead of
  submit) or suppressed where the reason is written down.
- Two findings are deliberate and excluded with the reason alongside: the
  client handing out its own facades, and the CNI conventional paths.

Coverage stood at 81.5%. New tests for image get/list/remove and pull failure,
Containers.inspect, event delivery and reconnection, JSON printing, the client
builder and NetworkAttachment bring it to 87.4%.

+++ src/main/java/io/nanofaas/containerd/internal/OciSpecBuilder.java
@@ -326,7 +326,8 @@ public final class OciSpecBuilder {
     }
 
     private static Value numberValue(long n) {
-        return Value.newBuilder().setNumberValue(n).build();
+        // Exact up to 2^53, far beyond any limit or id in a spec; JsonSupport prints it back as an integer.
+        return Value.newBuilder().setNumberValue((double) n).build();
     }
 
     private static Value boolValue(boolean b) {
